\name{shelters}
\alias{shelters}
\docType{data}
\title{
shelters
}
\description{
Pseudodata depicting the outcomes of 10,000 animals in shelters over three months generated to demonstrate the makeFlow() diagram. Aside: consider adopting your next pet :)
}
\usage{data("shelters")}
\format{
  A data frame with 10000 observations on the following 6 variables.
  \describe{
    \item{\code{loc}}{a factor with levels \code{shelter}}
    \item{\code{name}}{a character vector}
    \item{\code{id}}{a numeric vector}
    \item{\code{Jan}}{a factor with levels \code{Remaining} \code{Adopted} \code{Transfered} \code{Euthanized}}
    \item{\code{Feb}}{a factor with levels \code{Remaining} \code{Adopted} \code{Transfered} \code{Euthanized}}
    \item{\code{Mar}}{a factor with levels \code{Remaining} \code{Adopted} \code{Transfered} \code{Euthanized}}
  }
}

\keyword{datasets}
